Refactor test case, remove lexer unit tests
