+++ test/resources/export_expressions.ts
@@ -5,7 +5,6 @@ import { getSampleToken as t } from './resolved_tokens';
 
 interface TestCaseData {
   str: string;
-  tokens: Array<Token>;
   parseTree: Tree<Token>;
   ir: ExportExpression[];
   minimal_binary: Uint8Array;
@@ -13,7 +12,6 @@ interface TestCaseData {
 
 export const export_func_add_by_index: TestCaseData = {
   str: '(export "add" (func 0))',
-  tokens: ['(', 'export', '"add"', '(', 'func', '0', ')', ')'].map(t),
   parseTree: Tree.treeMap(['export', '"add"', ['func', '0']], t),
   ir: [new ExportExpression(t('"add"'), t('func'), t('0'))],
   minimal_binary: new Uint8Array([
