+++ test/resources/valid_function_bodies.ts
@@ -11,11 +11,10 @@ import { getSampleToken as t } from './resolved_tokens';
 
 export interface FunctionBodyTestCaseData {
   str: string;
-  tokens?: Array<Token>;
-  parseTree?: Tree<Token>;
-  ir?: IntermediateRepresentation;
-  unfolded_ir?: PureUnfoldedTokenExpression;
-  minimal_binary?: Uint8Array;
+  parseTree: Tree<Token>;
+  ir: IntermediateRepresentation;
+  unfolded_ir: PureUnfoldedTokenExpression;
+  minimal_binary: Uint8Array;
 }
 
 const simple_addition_sexpr: FunctionBodyTestCaseData = {
@@ -25,19 +24,6 @@ const simple_addition_sexpr: FunctionBodyTestCaseData = {
         (f64.const 1.5)
         )
     `,
-  tokens: [
-    '(',
-    'f64.add',
-    '(',
-    'f64.const',
-    '1',
-    ')',
-    '(',
-    'f64.const',
-    '1.5',
-    ')',
-    ')',
-  ].map(t),
   parseTree: [
     t('f64.add'),
     [t('f64.const'), t('1')],
@@ -61,7 +47,6 @@ const simple_addition_stack: FunctionBodyTestCaseData = {
         f64.const 1.5
         f64.add)
     `,
-  tokens: ['(', 'f64.const', '1', 'f64.const', '1.5', 'f64.add', ')'].map(t),
   parseTree: ['f64.const', '1', 'f64.const', '1.5', 'f64.add'].map(t),
   ir: new UnfoldedTokenExpression(
     ['f64.const', '1', 'f64.const', '1.5', 'f64.add'].map(t),
@@ -84,19 +69,6 @@ const nested_addition_stack: FunctionBodyTestCaseData = {
       f64.add
       f64.add
       `,
-  tokens: [
-    'f64.const',
-    '1',
-    'f64.const',
-    '1',
-    'f64.add',
-    'f64.const',
-    '1',
-    'f64.const',
-    '1',
-    'f64.add',
-    'f64.add',
-  ].map(t),
   parseTree: [
     'f64.const',
     '1',
@@ -196,18 +168,6 @@ const nested_addition_sexpr: FunctionBodyTestCaseData = {
       )
     )
       `,
-  tokens: [
-    ...['(', 'f64.add'],
-    ...['(', 'f64.add'],
-    ...['(', 'f64.const', '1', ')'],
-    ...['(', 'f64.const', '1', ')'],
-    ...[')'],
-    ...['(', 'f64.add'],
-    ...['(', 'f64.const', '1', ')'],
-    ...['(', 'f64.const', '1', ')'],
-    ')',
-    ')',
-  ].map(t),
   parseTree: [
     t('f64.add'),
     [t('f64.add'), [t('f64.const'), t('1')], [t('f64.const'), t('1')]],
@@ -281,26 +241,10 @@ const nested_addition_sexpr: FunctionBodyTestCaseData = {
   ]),
 };
 
-const allTestCases = [
+export const validTestCases = [
   simple_addition_sexpr,
   simple_addition_stack,
   nested_addition_stack,
   nested_addition_sexpr,
-  simple_addition_sexpr_without_argument_bracket_fails,
+  // simple_addition_sexpr_without_argument_bracket_fails,
 ];
-
-export const tokenizeTestCases = allTestCases.filter(
-  (testcase) => typeof testcase.tokens !== 'undefined',
-);
-export const parseTreeTestCases = allTestCases.filter(
-  (testcase) => typeof testcase.parseTree !== 'undefined',
-);
-export const irTestCases = allTestCases.filter(
-  (testcase) => typeof testcase.ir !== 'undefined',
-);
-export const unfoldIrTestCases = allTestCases.filter(
-  (testcase) => typeof testcase.unfolded_ir !== 'undefined',
-);
-export const minimalBinaryTestCases = allTestCases.filter(
-  (testcase) => typeof testcase.minimal_binary !== 'undefined',
-);
